Hackathon-PW/backend/pyproject.toml
[tool.poetry]
name = "hackathon-pw-backend"
version = "0.1.0"
description = "Algorand ASA / NFT backend"
authors = ["team"]

[tool.poetry.dependencies]
python = "^3.10"
algokit-utils = "2.4.0"
beaker-pyteal = "1.1.1"
pyteal = "0.24.1"
setuptools = "<81"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

